Move DMP files to a new folder
