+++ src/server.cpp
@@ -1,5 +1,5 @@
-#include "DMPRequest.hpp"
-#include "DMPResponse.hpp"
+#include "DMP/Request.hpp"
+#include "DMP/Response.hpp"
 
 int main () {
   std::cout << "HELLO SERVER DISTRIBUTED MOSQUETERS PROTOCOL" << std::endl;
